move EmbeddedBottomBar.tsx to pages/
