+++ gui/src/app/pages/EmbeddedBottomBar.tsx
